fix: refactor the release report library
so its easier to reuse in other plugins
